Move change listeners + filtering to addonManager
