+++ src/disco/addonManager.js
@@ -1,4 +1,15 @@
-import { installEventList } from 'disco/constants';
+import log from 'core/logger';
+
+import {
+  globalEvents,
+  installEventList,
+  ON_ENABLE,
+  ON_DISABLE,
+  ON_INSTALLING,
+  ON_UNINSTALLING,
+  ON_INSTALLED,
+  ON_UNINSTALLED,
+} from 'disco/constants';
 
 
 export function getAddon(guid, {_mozAddonManager = window.navigator.mozAddonManager} = {}) {
@@ -35,4 +46,43 @@ export function uninstall(guid, {_mozAddonManager = window.navigator.mozAddonMan
       }
       return;
     });
+}
+
+export function addChangeListeners(callback, mozAddonManager) {
+  function handleChangeEvent(e) {
+    const { id, type, needsRestart } = e;
+    const payload = { guid: id, needsRestart };
+    log.info('Event received', type, id, needsRestart);
+    switch (type) {
+      case 'onDisabled':
+        callback({type: ON_DISABLE, payload});
+        break;
+      case 'onEnabled':
+        callback({type: ON_ENABLE, payload});
+        break;
+      case 'onInstalling':
+        callback({type: ON_INSTALLING, payload});
+        break;
+      case 'onInstalled':
+        callback({type: ON_INSTALLED, payload});
+        break;
+      case 'onUninstalling':
+        callback({type: ON_UNINSTALLING, payload});
+        break;
+      case 'onUninstalled':
+        callback({type: ON_UNINSTALLED, payload});
+        break;
+      default:
+        throw new Error(`Unknown global event: ${type}`);
+    }
+  }
+
+  if (mozAddonManager && mozAddonManager.addEventListener) {
+    for (const event of globalEvents) {
+      mozAddonManager.addEventListener(event, handleChangeEvent);
+    }
+  } else {
+    log.info('mozAddonManager.addEventListener not available');
+  }
+  return handleChangeEvent;
 }
